+++ app/Main.hs
@@ -21,6 +21,7 @@ import Uniform.CmdLineArgs
 import UniformBase
 -- import           Data.Semigroup                 ( (<>) )
 import           Lib.ProcTxt
+import           Lib.ProcDirMD
 -- import           Lib.ProcPandocDatei
 import              Lib.OneMDfile
 import Lib.FileHandling
@@ -54,6 +55,7 @@ main = do
 data LitArgs = LitArgs 
     { isTxt   :: Bool   -- ^ is this a txt file
     , isDebug :: Bool   -- ^ switch to debug mode (unchanged file, NEW extension)
+    , isAllMD :: Bool   -- ^ collect all md files in directory
       , argfile  :: String -- ^ the filename absolute
       } deriving (Show)
 
@@ -67,10 +69,13 @@ cmdArgs =
     <*> switch 
             (long "debug" <> short 'd' <> help 
               "use debug mode; original file is unchange, new file with NEW extension attached")
+    <*> switch 
+            (long "all md" <> short 'm' <> help 
+              "process all md files in directory")
     <*> argument str
                  (
       --   long "filename" <>
-                  metavar "filename")
+                  metavar "filename or directory")
 
 parseAndExecute :: Text -> Text -> ErrIO ()
 parseAndExecute t1 t2 = do
@@ -79,16 +84,29 @@ parseAndExecute t1 t2 = do
     curr <- currentDir
     -- let dir0 = makeAbsDir "/home/frank/additionalSpace/DataBig/LitOriginals"
     let fn2     = argfile args :: FilePath
-    let fn = curr </> makeRelFile fn2 :: Path Abs File
     let isText  = isTxt args :: Bool
-
-    let ext     = getExtension fn
-    let isText2 = isText || ext == (Extension "txt") || ext == (Extension "md")
     let debug   = isDebug args :: Bool
+    let allMD   = isAllMD args :: Bool
+
     let fnErl =
             makeAbsFile "/home/frank/Workspace8/replaceUmlaut/nichtUmlaute.txt"
-    erl2         <- readErlaubt fnErl
-    if isText2 then procTxt debug erl2 fn else procMd1 debug erl2 fn
+    erl2  <- readErlaubt fnErl
+ 
+    if allMD 
+        then do
+            let dir = curr </> makeRelDir fn2 :: Path Abs Dir
+            procDirMD debug erl2 dir
+        else do
+            let fn = curr </> makeRelFile fn2 :: Path Abs File
+            let ext     = getExtension fn
+            let isText2 = isText || ext == (Extension "txt") || ext == (Extension "md")
+            if isText2 
+                then 
+                        procTxt debug erl2 fn 
+                else 
+                        procMd1 debug erl2 fn
+
+
 
     -- differences in md test for language and if nothing to do
         -- no new file is written