+++ src/Lib/FileHandling.hs
@@ -23,6 +23,14 @@ import Uniform.Pandoc (extMD)
 textlinesFile :: TypedFile5 Text [Text]
 -- txtFile = makeTyped txtExtension
 textlinesFile = makeTyped txtExtension
+textlinesNewFile :: TypedFile5 Text [Text]
+textlinesNewFile = makeTyped newExtension 
+textlinesBakFile :: TypedFile5 Text [Text]
+textlinesBakFile = makeTyped bakExtension 
+
+mdFile :: TypedFile5 Text [Text]
+mdFile = makeTyped extMD
+
 txtExtension :: Extension
 txtExtension = Extension "txt" :: Extension
 newExtension :: Extension
@@ -31,24 +39,21 @@ bakExtension :: Extension
 bakExtension = Extension "bak" :: Extension
 -- ^ filetype to read text in lines
 
-changeExtensionBakOrNew :: Bool -> Path Abs File -> ErrIO (Path Abs File)
--- ^ the given file is renamed as bak and the original filename returned
--- or (in casse of debug) the filename is changed with addition of 
--- NEW appended to filename (not extension)
-changeExtensionBakOrNew debug fn = if debug
-      then do
-          putIOwords ["changeExtensionBakOrNew  fn", showT fn]
-          let fnnew =   makeAbsFile ((removeExtension $ toFilePath fn) <> "NEW")
-          putIOwords ["changeExtensionBakOrNew fnnew", showT fnnew]
-          -- write8 fnnew markdownFileType ls4
-          -- putIOwords ["procMd result in new written", showT ls4, "fn", showT fnnew]
-          return fnnew
-      else do
-          let fnrename = fn <.> bakExtension :: Path Abs File
-          renameOneFile (fn <.> extMD) fnrename
-          putIOwords ["changeExtensionBakOrNew", 
-                        "procMd renamed to bak", showT fnrename]
-          return fn
+-- changeExtensionBakOrNew :: Bool -> Path Abs File -> ErrIO (Extension, Path Abs File, Path Abs File)
+-- -- ^ the given fn is split in extension and 
+-- -- returns the bak and the new file name  
+-- changeExtensionBakOrNew debug fn = do
+--     let fnNaked = makeAbsFile $ getParentDir fn </> getNakedFileName fn :: Path Abs File
+--     let fnExtension = getExtension fn :: Extension
+--     when debug $ putIOwords ["changeExtensionBakOrNew fnNaked", showT fnNaked, "fnExtension", showT fnExtension]
+ 
+--     let fnbak = fnNaked <.> bakExtension :: Path Abs File
+--     let fnnew = fnNaked <.> newExtension :: Path Abs File
+
+--     putIOwords ["changeExtensionBakOrNew fnbak", showT fnbak]
+--     putIOwords ["changeExtensionBakOrNew fnnew", showT fnnew]
+    
+--     return (fnExtension, fnbak, fnnew)
 
 instance TypedFiles7 Text [Text] where  -- creates sequence of lines
   wrap7 t = lines' t