+++ src/Lib/OneMDfile.hs
@@ -74,13 +74,13 @@ procMd1 debug erl2 fn = do
 
     let f1 =   mdDocRead f0 :: MdDoc1   -- definition in MdDocHandling
     let german = True -- mdocIsGerman f1  -- check requires pandoc?
-    when german $  do 
-            -- let f2 = updateMdDoc2 (procMdTxt erl2) (procMdTxt erl2) f1
-            newfn <- changeExtensionBakOrNew debug fn  -- not debug?
-            -- let f3 = mdDocWrite f2 
-            let f3 =  procTxt2 erl2   $ f0 -- process the header with it
-            writeFile2 newfn f3
-            when True $ putIOwords ["\n procMd1 ", showT fn, "german file umlaut changed with backup"]
+    -- when german $  do 
+    --         -- let f2 = updateMdDoc2 (procMdTxt erl2) (procMdTxt erl2) f1
+    --         newfn <- changeExtensionBakOrNew debug fn  -- not debug?
+    --         -- let f3 = mdDocWrite f2 
+    --         let f3 =  procTxt2 erl2   $ f0 -- process the header with it
+    --         writeFile2 newfn f3
+    --         when True $ putIOwords ["\n procMd1 ", showT fn, "german file umlaut changed with backup"]
             
 
     when debug $ putIOwords ["\n procMD1 ", showT fn, "file done with backup"]