+++ ChangeLog.md
@@ -1,2 +1,3 @@
 0.0.1.1 feb 2023 reconstruct work done fall 2022 - ghc 9.0.2
-0.0.1.2 feb reconstruct the freestanding replaceUmlaut
+         feb reconstruct the freestanding replaceUmlaut
+0.0.1.2 process all files in directory with md extension